Moved Maze class to a sep. file and minor fixes
