+++ CMakeLists.txt
@@ -26,8 +26,7 @@ endif()
 
 # Our Project
 
-add_executable(${PROJECT_NAME} main.cpp
-)
+add_executable(${PROJECT_NAME} main.cpp maze.cpp)
 #set(raylib_VERBOSE 1)
 target_link_libraries(${PROJECT_NAME} raylib)
